requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
construct==2.10.68
construct-typing==0.5.6
dnspython==2.6.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
jsonalias==0.1.1
loguru==0.7.2
magic-filter==1.0.12
motor==3.4.0
multidict==6.0.5
packaging==24.0
pluggy==1.4.0
psutil==5.9.8
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.3
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
solana==0.33.0
solders==0.21.0
Telethon==1.34.0
typing_extensions==4.10.0
urllib3==2.2.1
websockets==11.0.3
yarl==1.9.4
